Working card deletion (+ ran "flutter format" on src folder)

+++ lib/src/card_details_screen.dart
@@ -1,5 +1,6 @@
 // card_screen.dart
 
+import 'package:card_maker/src/deck_screen.dart';
 import 'package:flutter/material.dart';
 import 'package:card_maker/src/item_card.dart';
 import 'package:card_maker/src/utils/text_utils.dart';
@@ -23,38 +24,58 @@ class CardDetailsScreen extends StatelessWidget {
         actions: [
           IconButton(
             icon: const Icon(Icons.delete),
-            onPressed: () {
-              dataHandler.deleteItemCard(itemCard.id); 
-            },
+            onPressed: () => showDialog<String>(
+              context: context,
+              builder: (BuildContext context) => AlertDialog(
+                title: const Text("Delete?"),
+                content: const Text("Do you want to delete this item card?"),
+                actions: <Widget>[
+                  TextButton(
+                    onPressed: () => Navigator.pop(context, "No"),
+                    child: const Text("No"),
+                  ),
+                  TextButton(
+                    onPressed: () {
+                      dataHandler.deleteItemCard(itemCard.id);
+                      Navigator.push(
+                        context,
+                        MaterialPageRoute(
+                            builder: (context) => const DeckScreen()),
+                      );
+                    },
+                    child: const Text("Yes"),
+                  ),
+                ],
+              ),
+            ),
           )
         ],
       ),
-      body:
-        Column(
-          mainAxisAlignment: MainAxisAlignment.start,
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: <Widget>[
-            Padding(
-              padding: const EdgeInsets.only(
-                left: regPaddingSize, top: regPaddingSize, bottom: nrwPaddingSize
-              ),
-              child: cardHeaderText("Unique Item Card ID"),
-            ),
-            Padding(
+      body: Column(
+        mainAxisAlignment: MainAxisAlignment.start,
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: <Widget>[
+          Padding(
+            padding: const EdgeInsets.only(
+                left: regPaddingSize,
+                top: regPaddingSize,
+                bottom: nrwPaddingSize),
+            child: cardHeaderText("Unique Item Card ID"),
+          ),
+          Padding(
               padding: const EdgeInsets.only(left: regPaddingSize),
-              child: cardSubheaderText(itemCard.id)
-            ),
-            Padding(
-              padding: const EdgeInsets.only(
-                left: regPaddingSize, top: regPaddingSize, bottom: nrwPaddingSize
-              ),
-              child: cardHeaderText("Description"),
-            ),
-            Padding(
+              child: cardSubheaderText(itemCard.id)),
+          Padding(
+            padding: const EdgeInsets.only(
+                left: regPaddingSize,
+                top: regPaddingSize,
+                bottom: nrwPaddingSize),
+            child: cardHeaderText("Description"),
+          ),
+          Padding(
               padding: const EdgeInsets.only(left: regPaddingSize),
-              child: cardText(itemCard.desc)
-            )
-          ],
+              child: cardText(itemCard.desc))
+        ],
       ),
     );
   }
